refactor consts and reorder methods
