+++ client/src/components/InteractiveGrid/InteractiveGrid.jsx
@@ -2,69 +2,8 @@ import React, { Component } from 'react';
 import { DragDropContext } from 'react-beautiful-dnd';
 import { CourseList } from '..';
 import { Row, Col } from 'antd';
-
-const years = 5;
-const quarters = ["Fall", "Winter", "Spring", "Summer"];
-const numCoursePerQuarter = 4;
-
-const getItems = (years, quarters, numCoursePerQuarter) => {
-    let result = {};
-    for (let year = 0; year < years; year++) {
-        for (let quarter of quarters) {
-            result[`${year}-${quarter}`] = Array.from({ length: numCoursePerQuarter }, (v, k) => k).map(k => ({
-                id: `year-${year}-quarter-${quarter}-${k}`,
-                content: `year-${year}-quarter-${quarter}-${k}`,
-            }));
-        }
-    }
-    return result;
-};
-
-const reorder = (list, startIndex, endIndex) => {
-    const result = Array.from(list);
-    const [removed] = result.splice(startIndex, 1);
-    result.splice(endIndex, 0, removed);
-
-    return result;
-};
-
-const reorderCourses = ({ columns, source, destination }) => {
-    const current = [...columns[source.droppableId]];
-    const next = [...columns[destination.droppableId]];
-    const target = current[source.index];
-
-    let result = null;
-
-    // Tile was moved within the same list
-    if (source.droppableId === destination.droppableId) {
-        const reordered = reorder(
-            current,
-            source.index,
-            destination.index,
-        );
-        
-        result = {
-            ...columns,
-            [source.droppableId]: reordered,
-        };
-    } 
-    // Tile was moved to different list
-    else {
-        // Remove from original and insert into next
-        current.splice(source.index, 1);
-        next.splice(destination.index, 0, target);
-
-        result = {
-            ...columns,
-            [source.droppableId]: current,
-            [destination.droppableId]: next,
-        };
-    }
-
-    return {
-        columns: result
-    };
-};
+import { reorderCourses, getItems } from './utils';
+import { YEARS, QUARTERS, NUM_COURSES_PER_QUARTER } from './const';
 
 class InteractiveGrid extends Component {
     constructor(props) {
@@ -72,13 +11,11 @@ class InteractiveGrid extends Component {
         this.state = {
             major: "computer science",
             requirements: [],
-            columns: getItems(years, quarters, numCoursePerQuarter),
-            years,
-            quarters,
-            numCoursePerQuarter
+            columns: getItems(YEARS, QUARTERS, NUM_COURSES_PER_QUARTER),
+            years: YEARS,
+            quarters: QUARTERS,
+            numCoursePerQuarter: NUM_COURSES_PER_QUARTER
         };
-
-        // this.onDragEnd = this.onDragEnd.bind(this);
 
         props.getCoursesForMajor(this.state.major)
             .then(() => {
