+++ client/src/components/InteractiveGrid/const.js
@@ -1,0 +1,3 @@
+export const YEARS = 5;
+export const QUARTERS = ["Fall", "Winter", "Spring", "Summer"];
+export const NUM_COURSES_PER_QUARTER = 4;
